[Patchset 13] Refactor recordreader to not take the socket.

+++ src/profiling/memory/record_reader.cc
@@ -27,7 +27,8 @@ namespace perfetto {
 RecordReader::ReceiveBuffer RecordReader::BeginReceive() {
   if (read_idx_ < sizeof(record_size_))
     return {&record_size_, sizeof(record_size_) - read_idx_};
-  return {buf_.get(), (record_size_ + sizeof(record_size_) - read_idx_)};
+  return {buf_.get(), (static_cast<size_t>(record_size_) +
+                       sizeof(record_size_) - read_idx_)};
 }
 
 bool RecordReader::EndReceive(size_t recv_size, Record* record) {
